+++ src/Emby.StrmBridge/Properties/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Emby.StrmBridge.Tests")]
+[assembly: InternalsVisibleTo("Emby.StrmBridge.HostCheck")]